requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
atlas-provider-sqlalchemy
asyncpg
psycopg[binary]
python-multipart
watchfiles
bcrypt
websockets
google-generativeai